+++ tests/suite/to_kdl/inline.golden
@@ -1,0 +1,13 @@
+// MyStruct.new -(t: Type) (a: U60) (b: t) : (MyStruct t)
+ctr {MyStruct_new a b}
+
+// SadFn : U60
+fun (SadFn) {
+  (SadFn) =
+    let a0 = #2;
+    let a1 = #40;
+    let a2 = #400;
+    let a3 = #2000;
+    (+ (+ (+ a0 a1) a2) a3)
+}
+
