+++ tests/suite/to_kdl/inline_with_arg.golden
@@ -1,0 +1,1 @@
+You should not put arguments on the attribute 'inline'.
