+++ tests/suite/to_kdl/inline_with_arg.kind2
@@ -1,0 +1,4 @@
+#inline = Oi
+FunWithBadAttr : U60 {
+  1
+}
