+++ tests/suite/to_kdl/kdl_state_with_attr.golden
