+++ tests/suite/to_kdl/kdl_state_with_attr.kind2
