+++ tests/suite/to_kdl/non_inline_state.golden
@@ -1,1 +1,1 @@
-Function 'MyFn.state' must have exactly one rule with only variable patterns to be 'kdl_state'.
+Attribute 'kdl_state' requires 'MyFn.state' to have at least one rule.
